api/requirements.txt
pymongo==4.6.1
flask==3.0.0
mariadb==1.1.9
python-dotenv==1.0.1